feat: refactor API key management by creating dedicated service classes for generating, deleting, and updating API keys, and fetching usage analytics
