+++ src/controller/api/apiKeyController.ts
@@ -1,8 +1,5 @@
 import { FastifyInstance, FastifyRequest, FastifyReply } from "fastify";
 import { z } from "zod";
-import bcrypt from "bcrypt";
-import crypto from "crypto";
-import { prisma } from "../../utils/prisma";
 import { Response } from "../../types/responses";
 import { handleControllerError } from "../../utils/responseHelper";
 import { 
@@ -16,77 +13,20 @@ import {
 import { apiKeyGenerateSchema,
   type ApiKeyRequest
 } from "../../lib/api/generateApi/generateApiKey.schema";
-
-
-interface ApiKeyResponse {
-  apiKey: string;
-}
-
-interface DeleteApiKeyResponse {
-  message: string;
-}
-
-interface UpdateApiKeyResponse {
-  message: string;
-  apiKey: {
-    id: string;
-    name: string;
-    updatedAt: Date;
-  };
-}
-
-function generateApiKey(): string {
-  return `rdr_${crypto.randomBytes(32).toString('hex')}`;
-}
-
-function generateKeyName(): string {
-  return `API Key - ${new Date().toISOString()}`;
-}
+import { GenerateApiKeyService, type ApiKeyResponse } from "../../lib/api/generateApi/generateApiKey.service";
+import { DeleteApiKeyService, type DeleteApiKeyResponse } from "../../lib/api/deleteApiKey/deleteApiKey.service";
+import { UpdateApiKeyService, type UpdateApiKeyResponse } from "../../lib/api/updateApiKey/updateApiKey.service";
 export default async function apiKeyController(fastify: FastifyInstance) {
   fastify.post<{ Body: ApiKeyRequest }>(
     "/generate",
     async function (request: FastifyRequest<{ Body: ApiKeyRequest }>, reply: FastifyReply) {
       try {
         const validatedData = apiKeyGenerateSchema.parse(request.body);
-
-        const user = await prisma.user.findUnique({
-          where: {
-            email: validatedData.email
-          }
-        });
-
-        if (!user) {
-          const response: Response<ApiKeyResponse> = {
-            success: false,
-            error: "Invalid credentials"
-          };
-          return reply.code(401).send(response);
-        }
-
-        const isValidPassword = await bcrypt.compare(validatedData.password, user.password);
-
-        if (!isValidPassword) {
-          const response: Response<ApiKeyResponse> = {
-            success: false,
-            error: "Invalid credentials"
-          };
-          return reply.code(401).send(response);
-        }
-
-        const apiKey = generateApiKey();
-        const newApiKey = await prisma.apiKey.create({
-          data: {
-            key: apiKey,
-            name: generateKeyName(),
-            userId: user.id
-          }
-        });
+        const result = await GenerateApiKeyService.generateApiKey(validatedData);
 
         const response: Response<ApiKeyResponse> = {
           success: true,
-          data: {
-            apiKey: newApiKey.key
-          }
+          data: result
         };
         return reply.code(201).send(response);
       } catch (error) {
@@ -97,6 +37,15 @@ export default async function apiKeyController(fastify: FastifyInstance) {
           };
           return reply.code(400).send(response);
         }
+        
+        if (error instanceof Error && error.message === "Invalid credentials") {
+          const response: Response<ApiKeyResponse> = {
+            success: false,
+            error: error.message
+          };
+          return reply.code(401).send(response);
+        }
+
         handleControllerError(reply, error, "Internal server error");
         return;
       }
@@ -115,57 +64,11 @@ export default async function apiKeyController(fastify: FastifyInstance) {
           apiKeyId
         });
 
-        const user = await prisma.user.findUnique({
-          where: {
-            email: validatedData.email
-          }
-        });
-
-        if (!user) {
-          const response: Response<DeleteApiKeyResponse> = {
-            success: false,
-            error: "Invalid credentials"
-          };
-          return reply.code(401).send(response);
-        }
-
-        const isValidPassword = await bcrypt.compare(validatedData.password, user.password);
-
-        if (!isValidPassword) {
-          const response: Response<DeleteApiKeyResponse> = {
-            success: false,
-            error: "Invalid credentials"
-          };
-          return reply.code(401).send(response);
-        }
-
-        // Check if the API key exists and belongs to the user
-        const existingApiKey = await prisma.apiKey.findFirst({
-          where: {
-            id: apiKeyId,
-            userId: user.id
-          }
-        });
-
-        if (!existingApiKey) {
-          const response: Response<DeleteApiKeyResponse> = {
-            success: false,
-            error: "API key not found or access denied"
-          };
-          return reply.code(404).send(response);
-        }
-
-        await prisma.apiKey.delete({
-          where: {
-            id: apiKeyId
-          }
-        });
+        const result = await DeleteApiKeyService.deleteApiKey(validatedData);
 
         const response: Response<DeleteApiKeyResponse> = {
           success: true,
-          data: {
-            message: "API key deleted successfully"
-          }
+          data: result
         };
         return reply.code(200).send(response);
       } catch (error) {
@@ -176,6 +79,25 @@ export default async function apiKeyController(fastify: FastifyInstance) {
           };
           return reply.code(400).send(response);
         }
+
+        if (error instanceof Error) {
+          if (error.message === "Invalid credentials") {
+            const response: Response<DeleteApiKeyResponse> = {
+              success: false,
+              error: error.message
+            };
+            return reply.code(401).send(response);
+          }
+          
+          if (error.message === "API key not found or access denied") {
+            const response: Response<DeleteApiKeyResponse> = {
+              success: false,
+              error: error.message
+            };
+            return reply.code(404).send(response);
+          }
+        }
+
         handleControllerError(reply, error, "Internal server error");
         return;
       }
@@ -191,65 +113,11 @@ export default async function apiKeyController(fastify: FastifyInstance) {
         
         const validatedData = updateApiKeyRequestSchema.parse(request.body);
 
-        const user = await prisma.user.findUnique({
-          where: {
-            email: validatedData.email
-          }
-        });
-
-        if (!user) {
-          const response: Response<UpdateApiKeyResponse> = {
-            success: false,
-            error: "Invalid credentials"
-          };
-          return reply.code(401).send(response);
-        }
-
-        const isValidPassword = await bcrypt.compare(validatedData.password, user.password);
-
-        if (!isValidPassword) {
-          const response: Response<UpdateApiKeyResponse> = {
-            success: false,
-            error: "Invalid credentials"
-          };
-          return reply.code(401).send(response);
-        }
-
-        // Check if the API key exists and belongs to the user
-        const existingApiKey = await prisma.apiKey.findFirst({
-          where: {
-            id: apiKeyId,
-            userId: user.id
-          }
-        });
-
-        if (!existingApiKey) {
-          const response: Response<UpdateApiKeyResponse> = {
-            success: false,
-            error: "API key not found or access denied"
-          };
-          return reply.code(404).send(response);
-        }
-
-        const updatedApiKey = await prisma.apiKey.update({
-          where: {
-            id: apiKeyId
-          },
-          data: {
-            name: validatedData.name
-          }
-        });
+        const result = await UpdateApiKeyService.updateApiKey(validatedData, apiKeyId);
 
         const response: Response<UpdateApiKeyResponse> = {
           success: true,
-          data: {
-            message: "API key updated successfully",
-            apiKey: {
-              id: updatedApiKey.id,
-              name: updatedApiKey.name,
-              updatedAt: updatedApiKey.updatedAt
-            }
-          }
+          data: result
         };
         return reply.code(200).send(response);
       } catch (error) {
@@ -260,6 +128,25 @@ export default async function apiKeyController(fastify: FastifyInstance) {
           };
           return reply.code(400).send(response);
         }
+
+        if (error instanceof Error) {
+          if (error.message === "Invalid credentials") {
+            const response: Response<UpdateApiKeyResponse> = {
+              success: false,
+              error: error.message
+            };
+            return reply.code(401).send(response);
+          }
+          
+          if (error.message === "API key not found or access denied") {
+            const response: Response<UpdateApiKeyResponse> = {
+              success: false,
+              error: error.message
+            };
+            return reply.code(404).send(response);
+          }
+        }
+
         handleControllerError(reply, error, "Internal server error");
         return;
       }
