+++ src/controller/api/getApiKeysController.ts
@@ -1,7 +1,5 @@
 import { FastifyInstance, FastifyRequest, FastifyReply } from "fastify";
 import { z } from "zod";
-import bcrypt from "bcrypt";
-import { prisma } from "../../utils/prisma";
 import { Response } from "../../types/responses";
 import { handleControllerError } from "../../utils/responseHelper";
 import { 
@@ -12,50 +10,8 @@ import {
   getUsageAnalyticsRequestSchema,
   type GetUsageAnalyticsRequest
 } from "../../lib/api/getUsageAnalytics/getUsageAnalytics.schema";
-
-
-// Type for API key list response
-interface ApiKeyListResponse {
-  apiKeys: {
-    id: string;
-    key: string;
-    name: string;
-    createdAt: Date;
-    updatedAt: Date;
-    lastUsedAt: Date;
-    usageCount: number;
-    isActive: boolean;
-  }[];
-}
-
-// Type for usage analytics response
-interface UsageAnalyticsResponse {
-  totalApiKeys: number;
-  totalUsage: number;
-  activeApiKeys: number;
-  inactiveApiKeys: number;
-  mostUsedApiKey: {
-    id: string;
-    name: string;
-    usageCount: number;
-    lastUsedAt: Date;
-  } | null;
-  leastUsedApiKey: {
-    id: string;
-    name: string;
-    usageCount: number;
-    lastUsedAt: Date;
-  } | null;
-  averageUsagePerKey: number;
-  apiKeyDetails: Array<{
-    id: string;
-    name: string;
-    usageCount: number;
-    lastUsedAt: Date;
-    isActive: boolean;
-    createdAt: Date;
-  }>;
-}
+import { GetApiKeysService, type ApiKeyListResponse } from "../../lib/api/getApiKeys/getApiKeys.service";
+import { GetUsageAnalyticsService, type UsageAnalyticsResponse } from "../../lib/api/getUsageAnalytics/getUsageAnalytics.service";
 
 export default async function getApiKeysController(fastify: FastifyInstance) {
   fastify.post<{ Body: GetApiKeysRequest }>(
@@ -65,57 +21,12 @@ export default async function getApiKeysController(fastify: FastifyInstance) {
         // Validate the request body against the schema
         const validatedData = getApiKeysRequestSchema.parse(request.body);
 
-        // Find user by email
-        const user = await prisma.user.findUnique({
-          where: {
-            email: validatedData.email
-          }
-        });
-
-        // If user doesn't exist, return error
-        if (!user) {
-          const response: Response<ApiKeyListResponse> = {
-            success: false,
-            error: "Invalid credentials"
-          };
-          return reply.code(401).send(response);
-        }
-
-        // Compare password with hashed password
-        const isValidPassword = await bcrypt.compare(validatedData.password, user.password);
-
-        // If password is invalid, return error
-        if (!isValidPassword) {
-          const response: Response<ApiKeyListResponse> = {
-            success: false,
-            error: "Invalid credentials"
-          };
-          return reply.code(401).send(response);
-        }
-
-        // Fetch API keys for the user
-        const apiKeys = await prisma.apiKey.findMany({
-          where: {
-            userId: user.id
-          },
-          select: {
-            id: true,
-            key: true,
-            name: true,
-            createdAt: true,
-            updatedAt: true,
-            lastUsedAt: true,
-            usageCount: true,
-            isActive: true
-          }
-        });
+        const result = await GetApiKeysService.getApiKeys(validatedData);
 
         // Return the API keys
         const response: Response<ApiKeyListResponse> = {
           success: true,
-          data: {
-            apiKeys
-          }
+          data: result
         };
         return reply.code(200).send(response);
       } catch (error) {
@@ -126,6 +37,14 @@ export default async function getApiKeysController(fastify: FastifyInstance) {
             error: error.errors[0].message
           };
           return reply.code(400).send(response);
+        }
+
+        if (error instanceof Error && error.message === "Invalid credentials") {
+          const response: Response<ApiKeyListResponse> = {
+            success: false,
+            error: error.message
+          };
+          return reply.code(401).send(response);
         }
 
         handleControllerError(reply, error, "Internal server error");
@@ -141,99 +60,11 @@ export default async function getApiKeysController(fastify: FastifyInstance) {
       try {
         const validatedData = getUsageAnalyticsRequestSchema.parse(request.body);
 
-        // Find user by email
-        const user = await prisma.user.findUnique({
-          where: {
-            email: validatedData.email
-          }
-        });
-
-        if (!user) {
-          const response: Response<UsageAnalyticsResponse> = {
-            success: false,
-            error: "Invalid credentials"
-          };
-          return reply.code(401).send(response);
-        }
-
-        // Compare password with hashed password
-        const isValidPassword = await bcrypt.compare(validatedData.password, user.password);
-
-        if (!isValidPassword) {
-          const response: Response<UsageAnalyticsResponse> = {
-            success: false,
-            error: "Invalid credentials"
-          };
-          return reply.code(401).send(response);
-        }
-
-        // Build where clause for API keys
-        const whereClause: any = {
-          userId: user.id
-        };
-
-        // If specific API key ID is provided, filter by it
-        if (validatedData.apiKeyId) {
-          whereClause.id = validatedData.apiKeyId;
-        }
-
-        // Get all API keys for the user with usage data
-        const apiKeys = await prisma.apiKey.findMany({
-          where: whereClause,
-          select: {
-            id: true,
-            name: true,
-            usageCount: true,
-            lastUsedAt: true,
-            isActive: true,
-            createdAt: true
-          },
-          orderBy: {
-            usageCount: 'desc'
-          }
-        });
-
-        if (apiKeys.length === 0) {
-          const response: Response<UsageAnalyticsResponse> = {
-            success: false,
-            error: "No API keys found"
-          };
-          return reply.code(404).send(response);
-        }
-
-        // Calculate analytics
-        const totalApiKeys = apiKeys.length;
-        const totalUsage = apiKeys.reduce((sum, key) => sum + key.usageCount, 0);
-        const activeApiKeys = apiKeys.filter(key => key.isActive).length;
-        const inactiveApiKeys = totalApiKeys - activeApiKeys;
-        const averageUsagePerKey = totalUsage / totalApiKeys;
-
-        // Find most and least used API keys
-        const mostUsedApiKey = apiKeys[0]; // Already sorted by usage count desc
-        const leastUsedApiKey = apiKeys[apiKeys.length - 1];
+        const result = await GetUsageAnalyticsService.getUsageAnalytics(validatedData);
 
         const response: Response<UsageAnalyticsResponse> = {
           success: true,
-          data: {
-            totalApiKeys,
-            totalUsage,
-            activeApiKeys,
-            inactiveApiKeys,
-            mostUsedApiKey: {
-              id: mostUsedApiKey.id,
-              name: mostUsedApiKey.name,
-              usageCount: mostUsedApiKey.usageCount,
-              lastUsedAt: mostUsedApiKey.lastUsedAt
-            },
-            leastUsedApiKey: {
-              id: leastUsedApiKey.id,
-              name: leastUsedApiKey.name,
-              usageCount: leastUsedApiKey.usageCount,
-              lastUsedAt: leastUsedApiKey.lastUsedAt
-            },
-            averageUsagePerKey: Math.round(averageUsagePerKey * 100) / 100, // Round to 2 decimal places
-            apiKeyDetails: apiKeys
-          }
+          data: result
         };
 
         return reply.code(200).send(response);
@@ -244,6 +75,24 @@ export default async function getApiKeysController(fastify: FastifyInstance) {
             error: error.errors[0].message
           };
           return reply.code(400).send(response);
+        }
+
+        if (error instanceof Error) {
+          if (error.message === "Invalid credentials") {
+            const response: Response<UsageAnalyticsResponse> = {
+              success: false,
+              error: error.message
+            };
+            return reply.code(401).send(response);
+          }
+          
+          if (error.message === "No API keys found") {
+            const response: Response<UsageAnalyticsResponse> = {
+              success: false,
+              error: error.message
+            };
+            return reply.code(404).send(response);
+          }
         }
 
         handleControllerError(reply, error, "Internal server error");
