+++ src/lib/api/deleteApiKey/index.ts
@@ -1,0 +1,2 @@
+export * from './deleteApiKey.schema';
+export * from './deleteApiKey.service';
