+++ src/lib/api/generateApi/index.ts
@@ -1,0 +1,2 @@
+export * from './generateApiKey.schema';
+export * from './generateApiKey.service';
