+++ src/lib/api/getApiKeys/index.ts
@@ -1,0 +1,2 @@
+export * from './getApiKeys.schema';
+export * from './getApiKeys.service';
