+++ src/lib/api/getUsageAnalytics/index.ts
@@ -1,0 +1,2 @@
+export * from './getUsageAnalytics.schema';
+export * from './getUsageAnalytics.service';
