+++ src/lib/api/updateApiKey/index.ts
@@ -1,0 +1,2 @@
+export * from './updateApiKey.schema';
+export * from './updateApiKey.service';
